color coding for environments
the previous version of the right prompt didnt indicate the environment (staging/production/etc) that the cluster was in. Instead of wiritng it in text, lets indicate it with color.

The distinctive strings of the full environment names need to be present as enviornmental variables. One way to do it is to write them as exports in a file and then source it in profile.fish (this preventing staging domain names to be posted to github).